Other commit got corrupted in terminal.
I don't know how I'm going to even test this. I don't even know how I'm
going to make sure this works.

Tag or Elena might have a better idea on how this should integrate, but
of course the examples are not very detailed. So it's guesswork on all
sides.