I decided to swap the dynamic auto util with optional returns.
I thought it would be better to swap out the returns with an Optional
wrapper. This allows the caller to purposefully give a default value.
Which I think will be more helpful in the future than just letting this
default value return being left in and then hitting unexpected behavior
when you're debugging.